Move svg templates out of contracts folder
