+++ apps/dev-tool/next.config.js
@@ -13,17 +13,7 @@ const nextConfig = {
         {
           loader: "handlebars-loader",
           options: {
-            partialDirs: [
-              path.join(
-                __dirname,
-                "..",
-                "..",
-                "contracts",
-                "contracts",
-                "svg",
-                "partials"
-              ),
-            ],
+            partialDirs: [path.join(__dirname, "..", "..", "svg", "partials")],
             helperResolver(helper, callback) {
               if (helper.startsWith("./uint") || helper.startsWith("./int")) {
                 callback(
