+++ apps/minting-app/next.config.js
@@ -14,17 +14,7 @@ const nextConfig = {
         {
           loader: "handlebars-loader",
           options: {
-            partialDirs: [
-              path.join(
-                __dirname,
-                "..",
-                "..",
-                "contracts",
-                "contracts",
-                "svg",
-                "partials"
-              ),
-            ],
+            partialDirs: [path.join(__dirname, "..", "..", "svg", "partials")],
             helperResolver(helper, callback) {
               if (helper.startsWith("./uint") || helper.startsWith("./int")) {
                 callback(
