+++ contracts/contracts/interfaces/IWands.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.6;
 
 import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
-import "../svg/Template.sol";
+import "../Template.sol";
 
 interface IWands is IERC721 {
   struct Aspect {
