+++ contracts/package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "compile:template": "solid-mustache ./contracts/svg/template.svg.hbs -o ./contracts/svg/Template.sol --condense --dedupeThreshold=128",
+    "compile:template": "solid-mustache ../svg/template.svg.hbs -o ./contracts/Template.sol --condense --dedupeThreshold=128",
     "compile": "hardhat compile",
     "test": "hardhat test",
     "deploy": "hardhat deploy --network",
