+++ svg/partials/background.hbs
