+++ svg/partials/handles.hbs
