+++ svg/partials/sparkle.hbs
