+++ svg/partials/stars.hbs
