+++ svg/partials/stone.hbs
