+++ svg/partials/xpBar.hbs
