+++ svg/template.svg.hbs
